+++ Modules/SWMB/LocalMachine-Server.psm1
@@ -51,7 +51,7 @@ Function TweakDisablePasswordPolicy {
 	$tmpfile = New-TemporaryFile
 	secedit /export /cfg $tmpfile /quiet
 	(Get-Content $tmpfile).Replace("PasswordComplexity = 1", "PasswordComplexity = 0").Replace("MaximumPasswordAge = 42", "MaximumPasswordAge = -1") | Out-File $tmpfile
-	secedit /configure /db "$Env:SystemRoot\security\database\local.sdb" /cfg $tmpfile /areas SECURITYPOLICY | Out-Null
+	secedit /configure /db "${Env:SystemRoot}\security\database\local.sdb" /cfg $tmpfile /areas SECURITYPOLICY | Out-Null
 	Remove-Item -Path $tmpfile
 }
 
@@ -61,7 +61,7 @@ Function TweakEnablePasswordPolicy {
 	$tmpfile = New-TemporaryFile
 	secedit /export /cfg $tmpfile /quiet
 	(Get-Content $tmpfile).Replace("PasswordComplexity = 0", "PasswordComplexity = 1").Replace("MaximumPasswordAge = -1", "MaximumPasswordAge = 42") | Out-File $tmpfile
-	secedit /configure /db "$Env:SystemRoot\security\database\local.sdb" /cfg $tmpfile /areas SECURITYPOLICY | Out-Null
+	secedit /configure /db "${Env:SystemRoot}\security\database\local.sdb" /cfg $tmpfile /areas SECURITYPOLICY | Out-Null
 	Remove-Item -Path $tmpfile
 }